requirements.txt
hume==0.4.2
openai==1.16.2
pydub==0.25.1
python-dotenv==1.0.1
Requests==2.31.0
